requirements.txt
streamlit
jinja2
fpdf2
html5lib
openai